revno: 106 message:   dhis-web-maintenance-dataset: Fixed bug #342076 need swapping textarea
committer: Lars Helge Oeverland <larshelge@gmail.com>
branch nick: trunk
timestamp: Mon 2009-03-23 17:46:44 +0100

+++ dhis-2/dhis-web/dhis-web-maintenance/dhis-web-maintenance-dataset/src/main/java/org/hisp/dhis/dataset/action/section/AddSectionAction.java
@@ -200,12 +200,9 @@ public class AddSectionAction
         Section section = new Section();
 
         section.setName( sectionName );
-
         section.setLabel( sectionLabel );
-
         section.setSortOrder( 0 );
-
-        section.setDataSet( dataSetService.getDataSet( dataSetId.intValue() ) );
+        section.setDataSet( dataSetService.getDataSet( dataSetId ) );
 
         for ( String id : selectedList )
         {
